right where we started
