--- web-ext/.env.defaults
@@ -1,4 +1,0 @@
-# .env.defaults
-
-# Base URI for the server
-SERVER_BASE_URI=http://localhost:8787
