+++ web-ext/index.d.ts
@@ -3,13 +3,3 @@ declare module "*.jpg";
 declare module "*.jpeg";
 declare module "*.gif";
 declare module "*.svg";
-
-declare global {
-  namespace NodeJS {
-    interface ProcessEnv {
-      SERVER_BASE_URI: string;
-    }
-  }
-}
-
-export {};
