+++ web-ext/scripts/build.js
@@ -4,7 +4,6 @@ import tailwindcss from "tailwindcss";
 import autoprefixer from "autoprefixer";
 import CopyWebpackPlugin from "copy-webpack-plugin";
 import { CleanWebpackPlugin } from "clean-webpack-plugin";
-import DotEnvFlow from "dotenv-flow-webpack";
 
 export default (env) => {
   return {
@@ -128,7 +127,6 @@ export default (env) => {
           },
         ],
       }),
-      new DotEnvFlow(),
     ],
     resolve: {
       extensions: [".ts", ".tsx"],
